feat(#76): Phase 3 - Code Intelligence Features + Bug Fixes (#109)
Story #76: Phase 3 Code Intelligence Features + Bug Fixes - APPROVED

 MERGED BY CODEOWNER

Quality Gate Results:
- Tests: 150/150 passing 
- TypeScript: 0 errors 
- Acceptance Criteria: 9/9 met 
- Security: 15 patterns tested, 11 file types blocked 
- QA Status: APPROVED 

Features Implemented:
- Code extraction with metadata
- Sensitive data detection (15 patterns)
- File type blocking (11 credential types)
- Code insertion service
- Code suggestion panel with diff preview
- Command integration (3 new commands)
- TraceService config display fix
- StatisticsPanel labels update
- Conversation metadata tracking

Evidence: See PR #109 comments and .context/issues/76-phase-3-code-intelligence__gh/

This PR is production-ready and fully compliant with space_framework governance.
