+++ vscode-extension/v1/src/panels/TraceViewerPanel.ts
