+++ vscode-extension/v1/tests/services/TraceService.test.ts
